pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telecom-test-toolkit"
version = "1.0.0"
description = "A unified toolkit for telecom and 5G test engineers — log analysis, regression monitoring, flakiness detection, and report generation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "gbvk312", email = "gbvk312@users.noreply.github.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Logging",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=12.0",
    "jinja2>=3.0",
    "pandas>=1.5",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.20",
    "plotly>=5.0",
]
all = [
    "telecom-test-toolkit[dashboard]",
]

[project.scripts]
ttt = "ttt.cli:main"

[project.urls]
Homepage = "https://github.com/gbvk312/telecom-test-tools"
Repository = "https://github.com/gbvk312/telecom-test-tools"

[tool.setuptools.packages.find]
include = ["ttt*", "tools*"]
